Working vertex-frame assignment and rendering
Successfully can link vertices to frames, and can manipulate vertices by iterating through the vertex list in a frame. Vertex coordinates are local to their parent frame, but the class has a method to generate the global coordinates based on:
 - The vertex coordinates in the parent frame
 - The DCM of the parent frame
 - The origin coordinates of the parent frame in the global frame.
Wrote some basic code that uses the methods of these classes to render coordinate frames and points using matplotlib.